Always allow to trigger config change if decor insets change
This is restoring to previous behavior.

The timing to send config change when booting may slightly affect
the gpu memory usage of SystemUI. E.g. guess the usage may be a
bit less if the config change arrives before it has drawn.

And DisplayContent#sendNewConfiguration() has already checked
isReady(), so it is fine to let caller invoke directly.

Bug: 245170265
Test: android.platform.test.scenario.sleep.Idle
Change-Id: I7954925775e98171563ccad16b6999c17ece173c