Redraw dream overlay when a new dream is connected.
Now that we allow two dreams to have a breif overlap to have a smooth
transition, the overlay service can be bound to by two dreams at once.
Extras put in the intent are now moved to startDream because onBind()
is called only when the first dream is connected to it.

Also fix a bug that under an edge case where the service is destroyed
before dream is started, removing the view from window manager throws an
illegal argument exception.

Bug: 244315094
Bug: 249038780
Bug: 248496636
Fix: 244315094
Fix: 249038780
Test: atest DreamOverlayServiceTest
Test: manually on device, see video in bug

Change-Id: I8278e5b40a25539ce68a9c17c6cec9387d2da201