Add correct synchronization to VDMServiceTest.java
Many test cases in VirtualDeviceManagerServiceTest directly put elements
in mInputController.mInputDeviceDescriptors without acquiring the lock
of mInputController. In the context of the current test cases this is not
a problem (no parallel access occurs). However, the presubmit lint flags
this on new tests added. Therefore this commit adds synchronization to
all existing accesses.

Test: atest FrameworksServicesTests:VirtualDeviceManagerServiceTest
Bug: 244516415
Change-Id: Ib6b974e1aa672a6d7cd8152be8889a5107038a99