For non-udfps, detectFP if primaryAuth required
If the user attempts to FP auth, we will listen for FP
detection (as opposed to authentication or not at all). If the
user places their finger on the FP sensor, without running the
FP matcher, we can show the bouncer.

Additionally, update the logic when switching users. Immediately
after a user switch, cancel the FP listening state and then
restart based on the new state (because it may change to/from
detect or authenticate).

Repro steps:
  1. Enable extra logging:
       adb shell settings put global systemui/buffer/KeyguardUpdateMonitorLog v
  2. Setup side fps
  3. Fail side fps 5 times to lock out fp
  4. See logs that SysUI is running detectFP ("startListeningForFingerprint - detect")
  5. Attempt FP => bouncer appears

Test: atest KeyguardUpdateMonitorTest
Test: atest SystemUITests
Bug: 245778799
Change-Id: Ic1d4984c7514fc96b6552267128dee814730bb30
Merged-In: Ic1d4984c7514fc96b6552267128dee814730bb30