Describe requested visibilities in public types (1/n: client side)
We have been using InsetsVisibilities to store the requested visibility
of each internal type. But since the public caller cannot access the
internal insets types, we don’t need to store individual requested
visibilities of the internal types. We can just use an integer to
describe the requested visibilities of all public types. This is a step
of removing the internal insets types.

This CL is one of the many to achieve the task, which:

- defines "@InsetsType int mRequestedVisibleTypes" in InsetsController,
  (It will replace "InsetsVisibilities mRequestedVisibilities".)

- refactors the logic of updating/dispatching system UI visibility, and

- removes ITYPEs from DecorView.

Bug: 253420890
Bug: 234093736
Test: atest WindowInsetsControllerTests
Change-Id: Ib3d0eeb40db0943d150eec45b9c5a6df496da5d2