Check signature of updated font during boot
config_fontManagerServiceCert is a new config symbol that has
a list of der certificates to be used for font file verification.

After this change, the signature is stored next to the font file
and it is verified with the stored signature on boot time. If the
signature file is missing, e.g. due to device update, the updated
font file is discarded.

Bug: 242892591
Test: atest GtsFontHostTestCases
Test: atest UpdatableFontDirTest
Test: atest UpdatableSystemFontTest
Test: atest FontManagerTest
Test: Manually verified the font files can be updated
Change-Id: Ia4a6720b863163be523b845b883915185cee656c